Move all PingHub stuff to the main target
