+++ Podfile
@@ -35,6 +35,7 @@ abstract_target 'WordPress_Base' do
     pod 'Crashlytics'
     pod 'BuddyBuildSDK', '~> 1.0.11', :configurations => ['Release-Alpha']
     pod 'FLAnimatedImage', '~> 1.0'
+    pod 'Starscream', '~> 1.1.4' # Pinned to v1 for Swift 2
     # ----------------------------
     # Forked third party libraries
     # ----------------------------
@@ -70,12 +71,6 @@ abstract_target 'WordPress_Base' do
   target 'WordPressTodayWidget' do
   end
 
-end
-
-abstract_target :PingHub_Shared do
-  pod 'Starscream', '~> 1.1.4'
-  target :PingHub
-  target :PingHubTests
 end
 
 post_install do |installer_representation|
